Implement / Enforce API key usage - previously the key was hardcoded and never checked - minor version 0.9.0 reflects release candidate status - restructured api setup for readability - performance improvements when adding links
