+++ quasarr/storage/config.py
@@ -11,12 +11,15 @@ from Cryptodome.Cipher import AES
 from Cryptodome.Random import get_random_bytes
 from Cryptodome.Util.Padding import pad
 
-from quasarr.storage.sqlite_database import DataBase
 from quasarr.providers import shared_state
+from quasarr.storage.sqlite_database import DataBase
 
 
 class Config(object):
     _DEFAULT_CONFIG = {
+        'API': [
+            ("key", "secret", ""),
+        ],
         'JDownloader': [
             ("user", "secret", ""),
             ("password", "secret", ""),
